fix: revise 3917 - Count Indices With Opposite Parity - Easy

+++ src/problems/3901-4000/3917_count-indices-with-opposite-parity.js
@@ -23,7 +23,7 @@
  */
 const countOppositeParity = (nums) => {
         const n = nums.length;
-        const res = new Uint32Array(n);
+        const res = new Uint8Array(n);
         let cntOdd = 0;
         let cntEven = 0;
